feat: ts done
